refactor: split TrainNotes app into separate file
